+++ ImageProcessingForSwift/ViewController.swift
@@ -7,15 +7,6 @@
 
 import UIKit
 
-extension UIViewController {
-    static func loadFromNib() -> Self {
-        func instantiateFromNib<T: UIViewController>() -> T {
-            return T.init(nibName: String(describing: T.self), bundle: nil)
-        }
-        return instantiateFromNib()
-    }
-}
-
 final class ViewController: UIViewController {
     struct ImageProcessingData {
         let viewController: UIViewController
